template parameters count

Template types and parameters count matrix. The row names are template types. The column names are parameters short ids. The parameters are including the themes parameters and data column names. All the details are introduced in the full-page Excel file on GitHub.

Usage

template_parameters_count

Format

template_parameters_count

A data frame with template types and parameters 0/1 count matrix:

V1

head. file type head notice

V2

separator. select the separator which is used to delimit the data below (TAB,SPACE or COMMA).This separator must be used throughout this file.

V3

dataset name. label is used in the legend table

...